Just have LocationMap use Map directly
- These simplifications ought to just make it simpler to migrate to MapContentBuilder going forward.
